requirements.txt
streamlit==1.44.1
plotly==6.0.1
pypdf==5.4.0
google-generativeai>=0.5.0
pydantic>=2.0.0
